feat(manufacturing): Manufacturing/Production module (ADR-0035, V74-V76) + suite fixes
Manufacturing module: WorkOrder/WorkOrderComponent/WorkOrderOperation entities,
WO lifecycle (PLANNED→RELEASED→IN_PROGRESS→COMPLETED→CLOSED), BOM-pinned
component planning, WIP GL postings, costing service, variance report.

Integration fixes applied to merge this branch cleanly:
- AuditActions: restored all pre-manufacturing constants (linter had stripped them);
  appended WORKORDER_* section
- V75 manufacturing_coa_seed: account_name → name (column is named 'name' in V10)
- WorkOrderController: ApiResponse.created() → ApiResponse.ok() (no created() factory)
- WorkOrderCostingServiceImpl: WorkOrderCostReportDto constructor aligned to DTO signature
- WorkOrderServiceImplIT: fixed 8 stale call-site signatures against current DTOs
  (RequestContext.Principal, CreateUnitOfMeasureRequest, CreateProductRequest,
   ProductType.STORABLE→GOODS, CreateBomRequest, BomService.create, AddBomComponentRequest,
   ActivateBomRequest, SetOpeningValuationRequest pattern via stockService.openingBalance)
- EmailSender: @ConditionalOnBean(JavaMailSender.class) so context loads without SMTP
- NotificationRaiserImpl: Optional<EmailSender> injection (degrades cleanly, D-6)
- PostgresIntegrationTest: MailStubConfig inner @TestConfiguration (belt-and-suspenders)
- V69 budget_versions.version_no: SMALLINT → INTEGER (entity maps as int)

+++ backend/src/main/resources/db/migration/V69__budgets.sql
@@ -53,7 +53,7 @@ CREATE TABLE budget_versions (
     fiscal_year_id       BIGINT           NOT NULL,
     -- Denormalised for the single-active-version partial unique; NULL = company-wide
     cost_centre_value_id BIGINT,
-    version_no           SMALLINT         NOT NULL,
+    version_no           INTEGER          NOT NULL,
     status               VARCHAR(20)      NOT NULL DEFAULT 'DRAFT',
     label                VARCHAR(120),
     -- Self-FK: the version whose lines were copied as a seed (FR-BUD-08c)
